task2.2/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
open_clip_torch>=2.20.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
Pillow>=8.3.0
einops>=0.6.0
timm>=0.9.0
faiss-cpu>=1.7.0